+++ Calculator/src/main/java/ModelSub.java
@@ -4,17 +4,17 @@ public class ModelSub extends ModelCalc {
 
     }
     @Override
-    public Float result() {
+    public double result() {
         return x - y;
     }
 
     @Override
-    public void setX(Float value) {
+    public void setX(double value) {
         super.x = value;
     }
 
     @Override
-    public void setY(Float value) {
+    public void setY(double value) {
         super.y = value;
 
     }
